Initial commit
this is a template that i installed from online. its pure rust for ebpf,
kinda cool. anyway this project.

**ZING**

a high speed efficient ping server built on eBPF and rust.

**WHY**

fun and this is the first stage to learning eBPF. eBPF is probably good
for me to know for work and because i think its interesting. I have this
idea of making a distributed ping network and using eBPF to do most of
the IO.